refactor : cofrimbutton 컴포넌트를 /components 자리로 이동 , 기존에 있던 svg로 변경
