+++ src/lib/donationReward.ts
@@ -1,29 +1,29 @@
 export const rewardList = [
-    {
-      id: 1,
-      src: "/assets/icons/can.png",
-      alt: "can",
-      label: "밥 주기",
-      times: "2",
-      width: 40,
-      height: 40,
-    },
-    {
-      id: 2,
-      src: "/assets/icons/fish.svg",
-      alt: "fishing rod",
-      label: "놀아주기",
-      times: "2",
-      width: 40,
-      height: 40,
-    },
-    {
-      id: 3,
-      src: "/assets/icons/hand.png",
-      alt: "Hand",
-      label: "쓰다듬기",
-      times: "2",
-      width: 40,
-      height: 40,
-    },
-  ];
+  {
+    id: 1,
+    src: "/assets/icons/feed.svg",
+    alt: "can",
+    label: "밥 주기",
+    times: "2",
+    width: 54,
+    height: 50,
+  },
+  {
+    id: 2,
+    src: "/assets/icons/play.svg",
+    alt: "fishing rod",
+    label: "놀아주기",
+    times: "2",
+    width: 50,
+    height: 50,
+  },
+  {
+    id: 3,
+    src: "/assets/icons/stroke.svg",
+    alt: "Hand",
+    label: "쓰다듬기",
+    times: "2",
+    width: 50,
+    height: 50,
+  },
+];
